pyproject.toml
[project]
name = "chicago-crime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.3",
    "ipykernel>=7.1.0",
    "matplotlib>=3.10.8",
    "pandas>=3.0.0",
    "pyarrow>=23.0.0",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
    "sodapy>=2.2.0",
    "sqlalchemy>=2.0.46",
    "typing>=3.10.0.0",
]
